Add files via upload
A project to analyse how varying the cache parameters can influence the performance of a
processor to great extents. The architecture used for analysis is X86 architecture. Also to
note the best suited parameters for each benchmark through analysis of hit rates, miss rates
and CPIs.

458.sjeng.zip and 401_bzip2.zip contains all the text files extracted for the analysis
Project1_SPEC-master.zip contains the benchmarks
The command arguments used are mentioned in the Report.pdf